requirements.txt
discord.py==2.7.3
Flask==2.3.3
